form validation methods refactor
